Add defconfig for building ubilinux without i915
This bug: https://bugs.freedesktop.org/show_bug.cgi?id=106721
causes upboards to sometimes randomly hang and requires a hard
reboot by unplugging the power cable. This defconfig builds the
kernel without i915 while including all the pinctrl, LED and
FPGA MFD configs.

This is a config we use to test if the freezes are part of the i915
or the FPGA/CPLD chip

For some reason, building the kernel with i915 included and installing
it on an Upboard works. If the i915 is removed, the kernel fails to
boot as systemd fails to mount /tmp* stuff. No idea what this is, but
building with TMPFS enabled works.